media: video: tegra: nvavp: Fix build warning.
Cast values, so that printf is satisified.

Reviewed-on: http://review.adnet.avionic-design.de/234

+++ drivers/media/video/tegra/nvavp/nvavp_dev.c
@@ -912,8 +912,9 @@ static int nvavp_os_init(struct nvavp_info *nvavp)
 	}
 	dev_info(&nvavp->nvhost_dev->dev,
 		"using nvmem= carveout at %lx to load AVP os\n",
-		nvavp->os_info.phys);
-	sprintf(fw_os_file, "nvavp_os_%08lx.bin", nvavp->os_info.phys);
+		(unsigned long) nvavp->os_info.phys);
+	sprintf(fw_os_file, "nvavp_os_%08lx.bin",
+			(unsigned long) nvavp->os_info.phys);
 	nvavp->os_info.reset_addr = nvavp->os_info.phys;
 	nvavp->os_info.data = ioremap(nvavp->os_info.phys, SZ_1M);
 #endif
